requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
lxml==4.9.2
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
python-telegram-bot==20.0b0
requests==2.28.1
rfc3986==1.5.0
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
style==1.1.0
trio==0.22.0
trio-websocket==0.9.2
update==0.0.1
urllib3==1.26.13
wsproto==1.2.0
